- moves content type related tests to a dedicated file
